Implement generalized float accessor data decoding
This change introduces GLTFResourceReader::ReadFloatData that, in
contrast with ReadBinaryData, doesn't require knowledge of the specific
data type and instead decodes the data into floating-point values in
accordance with ES3 rules that are used by glTF specification for
animation and geometry data.

This makes the code much more concise, because there's no need to
enumerate individual types and implement custom fragile conversions, and
trivially extends the code to support additions to the geometry
encodings provided by KHR_mesh_quantization extension.