Add `GLTFSDK_API` definition to make calling convention modifier explicit (#137)
Without explicit calling convention modifers, consumers are required to compile with the same calling convention used in the released binaries. If a consumer compiles with the `__stdcall` calling convention, the linker won't resolve the glTF-SDK calls because the released binaries are built with the default `__cdecl` calling convention.

This change fixes the issue using an overridable `GLTFSDK_API` macro defined as `__cdecl` by default on Windows platforms or defined as empty on non-Windows platforms.